q_yaar_platform/requirements-dev.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
black==24.4.2
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
dj-database-url==2.2.0
Django==5.0.7
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-log-request-id==2.1.0
django-pgconnection==1.0.3
django-pghistory==3.2.0
django-pgtrigger==4.11.1
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
kombu==5.3.7
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.42
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13